Fix Import Issues for avr-gcc
Since this isn't a real STL implementation and only implements the
things I need, attempting to import these via `#include <utility>` has
never worked. I worked around this issue by appending the post-fix `.h`
in the import statement. This works when the code in question does not
also support systems with a standard library. The `either` library can
use this code when in an AVR environment, or use the real standard
library when in such an environment so this workaround no longer works.
This works around the problem by providing symbolic links without the
extensions